[18.0][OU-FIX] l10n_fr: let's upgrade l10n_fr_account

+++ openupgrade_scripts/scripts/l10n_fr_account/18.0.2.2/pre-migration.py
@@ -14,11 +14,9 @@ table_renames = [
 
 @openupgrade.migrate()
 def migrate(env, version):
-    openupgrade.rename_models(env.cr, model_renames)
-    openupgrade.rename_tables(env.cr, table_renames)
-    openupgrade.delete_records_safely_by_xml_id(
-        env,
-        [
-            "l10n_fr_fec.account_fr_fec_rule",
-        ],
-    )
+    if openupgrade.table_exists(env.cr, "account_fr_fec"):
+        openupgrade.rename_models(env.cr, model_renames)
+        openupgrade.rename_tables(env.cr, table_renames)
+        openupgrade.delete_records_safely_by_xml_id(
+            env, ["l10n_fr_fec.account_fr_fec_rule"]
+        )
